fix link and refactor
